Address changes in cask.el
Create copy of package-archives at startup to prevent cask.el nuking it; we
need it for pallet-init.

Allow cask-up to be sent a body to evaluate only if we have a Cask file,
and raise a message if we don't.

Amend pallet-repack to accept a t/nil arg and conditionally use the copy
of package-archives.

TODO: How to switch to the new .cask directory structure following
pallet-init?

TODO: Tests! Probably need to not mock cask.el.

TODO: Check we can still track manually installed/deleted packages.