Removes the need to use the Makefile while hacking on the project
This change needs the corresponding one-character change in el.mk.
This is crucial to avoid losing some time while working in `pallet.el`.
We should not have to use `make` in order to test every change.
If `@COMMENTARY` is not in an actual comment it is impossible to load `src/pallet.el` which slows down any workflow.